refactor(text_cursor): split monolith into focused modules with a thin facade
Break the ~1300-line text_cursor/__init__.py into cohesive modules
(constants, errors, models, uia, worker, discovery, ranges, snapshots,
operations, service) and reduce __init__.py to a small public facade.

- Single test seam: collaborators are called by module-level name and
  patched on their owning module; drop the dependency-injection params
  that duplicated that seam.
- Break the uia<->worker cycle: move find_caret_provider and
  try_get_caret_on_element into discovery.py so uia.py is a pure leaf.
- Trim the facade to the intended public API (I/O + action models,
  errors, run_tool); stop re-exporting internal plumbing.
- Remove dead code: the unused UIA_TEXT_PATTERN2_ID constant and
  worker.co_uninitialize.
- Minor cleanups: drop a redundant int() wrap and document
  SelectRelativeAction.origin.

Repoint tests to import internals from their owning submodules. No
behavior change; full suite passes.
