+++ src/windows_mcp/text_cursor/service.py
@@ -1,0 +1,116 @@
+"""Orchestrate TextCursor actions on the dedicated COM worker thread.
+
+Collaborators (find_caret_provider, make_snapshot, apply_write,
+snapshot_position, execute_sync) are referenced by their module-level names so
+that tests can substitute them with monkeypatch.setattr on this module.
+"""
+
+from __future__ import annotations
+
+import asyncio
+
+from .errors import TextCursorVerificationError
+from .models import CursorAction, CursorToolResult, GetInfoAction
+from .discovery import find_caret_provider
+from .operations import WriteAction, apply_write
+from .snapshots import make_snapshot, snapshot_position
+from .worker import EXECUTOR
+
+
+def run_get_info(action: GetInfoAction) -> CursorToolResult:
+    """Read information about the focused caret or selection."""
+    caret_info = find_caret_provider()
+    snapshot = make_snapshot(caret_info, action.context_chars)
+
+    return CursorToolResult(
+        success=True,
+        mode=action.mode,
+        message="Caret information acquired.",
+        after=snapshot,
+        warnings=snapshot.warnings,
+    )
+
+
+def run_write(action: WriteAction) -> CursorToolResult:
+    """Apply a write action and report the resulting caret or selection."""
+    caret_info = find_caret_provider()
+
+    before = make_snapshot(caret_info, action.context_chars)
+    verified, target = apply_write(action, caret_info)
+    requested = action.model_dump(
+        exclude={
+            "delay",
+            "context_chars",
+            "verify",
+        }
+    )
+
+    # Always reacquire the focused provider before reporting success or a
+    # verification mismatch. The values returned by TextRange.Move describe
+    # only the client-side target; `actual` must come from the real provider.
+    refreshed = find_caret_provider()
+    after = make_snapshot(refreshed, action.context_chars)
+    actual = snapshot_position(after)
+
+    if verified is False:
+        raise TextCursorVerificationError(
+            "The provider accepted the operation, but read-back verification "
+            "showed that the real caret/selection did not match the calculated "
+            f"target. Requested: {requested}; target: {target}; actual: {actual}."
+        )
+
+    warnings = list(dict.fromkeys([*before.warnings, *after.warnings]))
+
+    return CursorToolResult(
+        success=True,
+        mode=action.mode,
+        message="Operation applied.",
+        verified=verified,
+        requested=requested,
+        target=target,
+        actual=actual,
+        before=before,
+        after=after,
+        warnings=warnings,
+    )
+
+
+def execute_sync(action: CursorAction) -> CursorToolResult:
+    """Execute one action on the dedicated COM worker thread.
+
+    COM is initialized once per worker thread and the UIA client is cached on
+    first use; there is no per-call CoInitialize/CoUninitialize or CreateObject.
+    """
+    if isinstance(action, GetInfoAction):
+        return run_get_info(action)
+
+    return run_write(action)
+
+
+async def run_tool(action: CursorAction) -> CursorToolResult:
+    """
+    Inspect or manipulate the focused Windows text control through UIA.
+    Modes:
+    - get_info
+    - move_relative
+    - move_absolute
+    - select_relative
+    - select_absolute
+    - select_all
+    - collapse_selection
+    Every mode accepts `delay`, expressed in seconds. The delay occurs before
+    the focused UIA element is located, so the caller can focus the target
+    control during that interval.
+    Absolute move/select inputs and returned offsets both use provider-defined
+    UIA TextUnit_Character steps from DocumentRange start. Returned offsets can
+    be passed directly to absolute move/select actions.
+    """
+    if action.delay > 0:
+        await asyncio.sleep(action.delay)
+
+    loop = asyncio.get_running_loop()
+    return await loop.run_in_executor(
+        EXECUTOR,
+        execute_sync,
+        action,
+    )
